[Bugfix][DSV4] Use project-native streaming tool_choice="none" patch
Rename the streaming tool_choice="none" guard to
patch_dsv4_dsml_stream_consistent.py (switching the platform __init__ import
accordingly). Same minimal approach: wrap the public
DelegatingParser.extract_tool_calls_streaming so that when tool_choice is
None/"none" the delta text is surfaced as plain DeltaMessage.content instead of
parsed into tool_calls; DeltaMessage is imported from
vllm.entrypoints.openai.engine.protocol (what abstract_parser uses on vLLM
v0.20.2). Streaming matches the non-streaming none path (markup stays in
content, finish_reason="stop"); reasoning and auto/forced paths untouched.
